cleaned up controller, moved mapping to userService

+++ UserApi/Data/UserRepository.cs
@@ -31,6 +31,12 @@ namespace UserApi.Data
         public async Task<User> DeleteAsync(int id)
         {
             var user = await _context.Users.FindAsync(id);
+
+            if(user == null)
+            {
+                return null;
+            }
+
             _context.Users.Remove(user);
             await _context.SaveChangesAsync();
             return user;
